refactor: consolidate flip_logging (6→3 files) and fix trust service tests
Library simplification:
- Inline context.py (single ContextVar) into formatter.py
- Merge filters.py (PIIRedactionFilter) into formatter.py
- Inline 3 used event constants into middleware.py, delete events.py
- Public API unchanged

Test fixes:
- Add ../../libs to PYTHONPATH and pytest pythonpath so flip_logging is
  importable without Docker volume mounts
- Add conftest.py to each trust service with dummy env vars for Settings()
- All 93 tests now pass across all 3 services (was 0 passing before)
