--- libs/flip_logging/context.py
@@ -1,17 +1,0 @@
-# Copyright (c) Guy's and St Thomas' NHS Foundation Trust & King's College London
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#     http://www.apache.org/licenses/LICENSE-2.0
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-from contextvars import ContextVar
-
-# Per-request context: holds a dict of fields (request_id, user_id, etc.)
-# that get merged into every log record emitted during that request.
-request_context: ContextVar[dict | None] = ContextVar("request_context", default=None)
