+++ trust/data-access-api/pyproject.toml
@@ -40,7 +40,7 @@ dev = [
 ]
 
 [tool.pytest.ini_options]
-pythonpath = "data_access_api"
+pythonpath = ["data_access_api", "../../libs"]
 python_files = ["test_*.py", "*_test.py"]
 addopts = []
 filterwarnings = ["ignore::DeprecationWarning", "ignore::FutureWarning"]
