+++ trust/imaging-api/.env
@@ -10,4 +10,4 @@
 # limitations under the License.
 #
 
-PYTHONPATH=./imaging_api
+PYTHONPATH=./imaging_api:../../libs
