+++ trust/imaging-api/pyproject.toml
@@ -43,7 +43,7 @@ dev = [
 ]
 
 [tool.pytest.ini_options]
-pythonpath = "imaging_api"
+pythonpath = ["imaging_api", "../../libs"]
 python_files = ["test_*.py", "*_test.py"]
 addopts = []
 filterwarnings = ["ignore::DeprecationWarning", "ignore::FutureWarning"]
