+++ trust/trust-api/.env
@@ -10,4 +10,4 @@
 # limitations under the License.
 #
 
-PYTHONPATH=./trust_api
+PYTHONPATH=./trust_api:../../libs
