+++ trust/trust-api/pyproject.toml
@@ -36,7 +36,7 @@ dev = [
 ]
 
 [tool.pytest.ini_options]
-pythonpath = "trust_api"
+pythonpath = ["trust_api", "../../libs"]
 python_files = ["test_*.py", "*_test.py"]
 addopts = []
 filterwarnings = ["ignore::DeprecationWarning", "ignore::FutureWarning"]
