+++ trust/trust-api/tests/conftest.py
@@ -10,4 +10,18 @@
 # limitations under the License.
 #
 
-PYTHONPATH=./data_access_api:../../libs
+import os
+
+# Set dummy environment variables required by Settings() before any app code
+# is imported.  These are only used in tests; real values come from Docker
+# Compose environment or .env files in deployed environments.
+_TEST_ENV_DEFAULTS = {
+    "CENTRAL_HUB_API_URL": "http://localhost:8000",
+    "DATA_ACCESS_API_URL": "http://localhost:8001",
+    "IMAGING_API_URL": "http://localhost:8002",
+    "PRIVATE_API_KEY": "test-key",
+    "PRIVATE_API_KEY_HEADER": "Authorization",
+}
+
+for key, value in _TEST_ENV_DEFAULTS.items():
+    os.environ.setdefault(key, value)
